ZIM-FOOD/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
requests==2.31.0
supabase==2.3.1
mistralai==0.0.9
python-multipart==0.0.9
httpx==0.27.0
gunicorn==21.2.0
